WIP #269 restructure frame_grabber
- code reduction from 1.6M/141K/1.5M
  to 1.55M/128K/1.49M
- use node frame_grabber as default, replace it in webpack.config.js with the browser frame grabber when building for browser.
