+++ configs/webpack.config.js
@@ -46,6 +46,7 @@ module.exports = {
             ENV: require(path.join(__dirname, './env/', process.env.BUILD_ENV)),
         }),
         new webpack.NormalModuleReplacementPlugin(/input_stream_node/, './input_stream/input_stream_browser'),
+        new webpack.NormalModuleReplacementPlugin(/..\/input\/frame_grabber/, '../input/frame_grabber_browser'),
     ],
     optimization: {
         minimize: false,
