+++ src/input/frame_grabber.js
@@ -1,54 +1,37 @@
-import NodeFrameGrabber from './frame_grabber_node';
+// NOTE FOR ANYONE IN HERE IN THE FUTURE: This module is used when the module is built for use in Node.
+// Webpack.config.js explicitly REPLACES this module with the file called frame_grabber_browser when it is packing the Browser distribution.
 
-import {
-    imageRef,
-    grayAndHalfSampleFromCanvasData,
-    computeGray,
-} from '../common/cv_utils';
-
-const TO_RADIANS = Math.PI / 180;
-
-function adjustCanvasSize(canvas, targetSize) {
-    if (canvas.width !== targetSize.x) {
-        if (ENV.development) {
-            console.log('WARNING: canvas-size needs to be adjusted');
-        }
-        canvas.width = targetSize.x;
-    }
-    if (canvas.height !== targetSize.y) {
-        if (ENV.development) {
-            console.log('WARNING: canvas-size needs to be adjusted');
-        }
-        canvas.height = targetSize.y;
-    }
-}
+const CVUtils = require('../common/cv_utils');
+const Ndarray = require('ndarray');
+const Interp2D = require('ndarray-linear-interpolate').d2;
 
 const FrameGrabber = {};
 
 FrameGrabber.create = function (inputStream, canvas) {
     const _that = {};
-    const _streamConfig = inputStream.getConfig();
-    const _videoSize = imageRef(inputStream.getRealWidth(), inputStream.getRealHeight());
+    const _videoSize = CVUtils.imageRef(inputStream.getRealWidth(), inputStream.getRealHeight());
     const _canvasSize = inputStream.getCanvasSize();
-    const _size = imageRef(inputStream.getWidth(), inputStream.getHeight());
-    const topRight = inputStream.getTopRight();
-    const _sx = topRight.x;
-    const _sy = topRight.y;
-    let _canvas;
-    let _ctx = null;
-    let _data = null;
+    const _size = CVUtils.imageRef(inputStream.getWidth(), inputStream.getHeight());
+    const _topRight = inputStream.getTopRight();
+    let _data = new Uint8Array(_size.x * _size.y);
+    const _grayData = new Uint8Array(_videoSize.x * _videoSize.y);
+    const _canvasData = new Uint8Array(_canvasSize.x * _canvasSize.y);
+    /* eslint-disable new-cap */
+    const _grayImageArray = Ndarray(_grayData, [_videoSize.y, _videoSize.x]).transpose(1, 0);
+    const _canvasImageArray = Ndarray(_canvasData, [_canvasSize.y, _canvasSize.x]).transpose(1, 0);
+    const _targetImageArray = _canvasImageArray
+        .hi(_topRight.x + _size.x, _topRight.y + _size.y)
+        .lo(_topRight.x, _topRight.y);
+    const _stepSizeX = _videoSize.x / _canvasSize.x;
+    const _stepSizeY = _videoSize.y / _canvasSize.y;
 
-    _canvas = canvas || document.createElement('canvas');
-    _canvas.width = _canvasSize.x;
-    _canvas.height = _canvasSize.y;
-    _ctx = _canvas.getContext('2d');
-    _data = new Uint8Array(_size.x * _size.y);
     if (ENV.development) {
         console.log('FrameGrabber', JSON.stringify({
-            size: _size,
-            topRight,
-            videoSize: _videoSize,
-            canvasSize: _canvasSize,
+            videoSize: _grayImageArray.shape,
+            canvasSize: _canvasImageArray.shape,
+            stepSize: [_stepSizeX, _stepSizeY],
+            size: _targetImageArray.shape,
+            topRight: _topRight,
         }));
     }
 
@@ -71,46 +54,40 @@ FrameGrabber.create = function (inputStream, canvas) {
      * The image-data is converted to gray-scale and then half-sampled if configured.
      */
     _that.grab = function () {
-        const doHalfSample = _streamConfig.halfSample;
         const frame = inputStream.getFrame();
-        let drawable = frame;
-        let drawAngle = 0;
-        let ctxData;
-        if (drawable) {
-            adjustCanvasSize(_canvas, _canvasSize);
-            if (_streamConfig.type === 'ImageStream') {
-                drawable = frame.img;
-                if (frame.tags && frame.tags.orientation) {
-                    switch (frame.tags.orientation) {
-                        case 6:
-                            drawAngle = 90 * TO_RADIANS;
-                            break;
-                        case 8:
-                            drawAngle = -90 * TO_RADIANS;
-                            break;
-                    }
-                }
-            }
 
-            if (drawAngle !== 0) {
-                _ctx.translate(_canvasSize.x / 2, _canvasSize.y / 2);
-                _ctx.rotate(drawAngle);
-                _ctx.drawImage(drawable, -_canvasSize.y / 2, -_canvasSize.x / 2, _canvasSize.y, _canvasSize.x);
-                _ctx.rotate(-drawAngle);
-                _ctx.translate(-_canvasSize.x / 2, -_canvasSize.y / 2);
-            } else {
-                _ctx.drawImage(drawable, 0, 0, _canvasSize.x, _canvasSize.y);
-            }
-
-            ctxData = _ctx.getImageData(_sx, _sy, _size.x, _size.y).data;
-            if (doHalfSample) {
-                grayAndHalfSampleFromCanvasData(ctxData, _size, _data);
-            } else {
-                computeGray(ctxData, _data, _streamConfig);
-            }
+        if (frame) {
+            this.scaleAndCrop(frame);
             return true;
         }
         return false;
+    };
+
+    // eslint-disable-next-line
+    _that.scaleAndCrop = function(frame) {
+        // 1. compute full-sized gray image
+        CVUtils.computeGray(frame.data, _grayData);
+
+        // 2. interpolate
+        for (let y = 0; y < _canvasSize.y; y++) {
+            for (let x = 0; x < _canvasSize.x; x++) {
+                // eslint-disable-next-line no-bitwise
+                _canvasImageArray.set(x, y, (Interp2D(_grayImageArray, x * _stepSizeX, y * _stepSizeY)) | 0);
+            }
+        }
+
+        // targetImageArray must be equal to targetSize
+        if (_targetImageArray.shape[0] !== _size.x
+            || _targetImageArray.shape[1] !== _size.y) {
+            throw new Error('Shapes do not match!');
+        }
+
+        // 3. crop
+        for (let y = 0; y < _size.y; y++) {
+            for (let x = 0; x < _size.x; x++) {
+                _data[y * _size.x + x] = _targetImageArray.get(x, y);
+            }
+        }
     };
 
     _that.getSize = function () {
@@ -120,5 +97,4 @@ FrameGrabber.create = function (inputStream, canvas) {
     return _that;
 };
 
-export default FrameGrabber;
-export { NodeFrameGrabber };
+module.exports = FrameGrabber;
